Update Home.js
I suggest using BogSwap since it aggregates both SpookySwap and SpiritSwap providing the best route with no extra fees. 

There is an easy way to make it an embed in a popup so you users won't need to leave the website. Just add &embed=1 at the end of URL. 

Feel free to DM in Telegram @bscden